ci-cd/ci_cd_resources/requirements.txt
aws_cdk.aws_codebuild
aws-cdk.core
boto3
